move and rename files to correct places +
