+++ apps/bot/src/discord/utils/build-menu-buttons.ts
@@ -2,7 +2,7 @@ import type { ButtonBuilder } from 'discord.js';
 import { ActionRowBuilder } from 'discord.js';
 
 import { CREW_BUTTON_NAME } from '../../domains/crew/constants.js';
-import { buildCrewNavButton } from '../../domains/crew/crew-nav-button.js';
+import { buildCrewNavButton } from '../../domains/crew/utils/build-crew-nav-button.js';
 import { PROFIL_BUTTON_NAME } from '../../domains/player/constants.js';
 import { buildProfilButton } from '../../domains/player/profil-button.js';
 import { INVENTORY_BUTTON_NAME } from '../../domains/resource/constants.js';
