+++ apps/bot/src/domains/crew/utils/build-crew-nav-button.ts
@@ -1,9 +1,8 @@
 import { ButtonBuilder, ButtonStyle } from 'discord.js';
 
-import { buildCustomId } from '../../discord/utils/index.js';
-import { DOMAIN_EMOJI } from '../../shared/domains.js';
-
-import { CREW_BUTTON_NAME } from './constants.js';
+import { buildCustomId } from '../../../discord/utils/index.js';
+import { DOMAIN_EMOJI } from '../../../shared/domains.js';
+import { CREW_BUTTON_NAME } from '../constants.js';
 
 const FIRST_PAGE = 0;
