+++ apps/bot/src/domains/crew/utils/build-crew-view.ts
@@ -1,12 +1,12 @@
 import type { Player, Ship } from '@one-piece/db';
 
-import type { View } from '../../discord/types.js';
-import { buildMenuButtons, buildOpEmbed, buildPaginationButtons, clampPage, splitIntoPages } from '../../discord/utils/index.js';
-import type { CharacterRow } from '../character/types.js';
-import { getCharacterInstanceName } from '../character/utils/index.js';
-import { getCrewCapacity } from '../crew/capacity.js';
+import type { View } from '../../../discord/types.js';
+import { buildMenuButtons, buildOpEmbed, buildPaginationButtons, clampPage, splitIntoPages } from '../../../discord/utils/index.js';
+import type { CharacterRow } from '../../character/types.js';
+import { getCharacterInstanceName } from '../../character/utils/index.js';
+import { CREW_BUTTON_NAME } from '../constants.js';
 
-import { CREW_BUTTON_NAME } from './constants.js';
+import { getCrewCapacity } from './get-crew-capacity.js';
 
 export function buildCrewView(player: Player, ship: Ship, characters: Array<CharacterRow>, page: number): View {
   const menuRow = buildMenuButtons(CREW_BUTTON_NAME, player.id);
