+++ apps/bot/src/domains/crew/utils/get-crew-capacity.ts
